requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
tflite-runtime==2.14.0
insightface==0.7.3
onnxruntime==1.17.3
qdrant-client==1.8.2
python-multipart==0.0.9
prometheus-client==0.20.0
httpx==0.27.0
paramiko==3.4.0
pyyaml==6.0.1